+++ src/components/ErrorView.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@/components/ui/button";
+import { Button } from "@/shared/components/ui/button";
 
 const ErrorIcon = () => (
 	<svg
