+++ src/components/LoadingView.tsx
@@ -1,6 +1,6 @@
-import { Progress } from "@/components/ui/progress";
-import { Card, CardContent } from "@/components/ui/card";
-import { Button } from "@/components/ui/button";
+import { Progress } from "@/shared/components/ui/progress";
+import { Card, CardContent } from "@/shared/components/ui/card";
+import { Button } from "@/shared/components/ui/button";
 import { CheckCircle2, Loader2, CircleDashed } from "lucide-react";
 import type { Step } from "@/api/audienceProgress";
